test(order-resolution): テストの整理
