+++ server/src/domain/phase.rs
@@ -364,3 +364,16 @@ impl PhaseContext {
 /// フェイズの終了結果（必要に応じてフィールドを追加）
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
 pub struct PhaseCloseResult {}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_new_spring_order_always_increments_year() {
+        let p = Phase::new_spring_order(1900, 7);
+        assert_eq!(p.year(), 1901);
+        assert_eq!(p.index(), 8);
+        assert!(matches!(p.phase_type(), PhaseKind::SpringOrder(_)));
+    }
+}
